chore: refactor sqlite query to support subscribe
