+++ crates/common/src/db.rs
@@ -1,7 +1,6 @@
 use std::str::FromStr;
 
 use eyre::Result;
-use reth_primitives::B256;
 use reth_tracing::tracing::debug;
 use sqlx::{
     sqlite::{SqliteConnectOptions, SqlitePoolOptions},
