+++ crates/rpc/src/apis/types.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[serde(untagged)]
+pub enum AddressRepresentation {
+    ArrayOfStrings(Vec<String>),
+    Bytes([u8; 20]),
+    String(String),
+}
